abstract test base
