+++ nullaway/src/test/java/com/uber/nullaway/NullAwayEnsuresNonNullTests.java
@@ -1,14 +1,8 @@
 package com.uber.nullaway;
 
-import com.google.errorprone.CompilationTestHelper;
-import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.TemporaryFolder;
 
-public class NullAwayEnsuresNonNullTests {
-  @Rule public final TemporaryFolder temporaryFolder = new TemporaryFolder();
-
-  private CompilationTestHelper defaultCompilationHelper;
+public class NullAwayEnsuresNonNullTests extends NullAwayTestsBase {
 
   @Test
   public void requiresNonNullInterpretation() {
