+++ nullaway/src/test/java/com/uber/nullaway/NullAwayThriftTests.java
@@ -1,15 +1,8 @@
 package com.uber.nullaway;
 
-import com.google.errorprone.CompilationTestHelper;
-import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.TemporaryFolder;
 
-public class NullAwayThriftTests {
-  @Rule public final TemporaryFolder temporaryFolder = new TemporaryFolder();
-
-  private CompilationTestHelper defaultCompilationHelper;
-
+public class NullAwayThriftTests extends NullAwayTestsBase {
   @Test
   public void testThriftIsSet() {
     defaultCompilationHelper
